ihome/requirement.txt
flask
flask-script
flask-blueprint
flask-sqlalchemy
flask-login
pymysql
flask-session
redis

